replace jest with tap
