+++ package.json
@@ -12,8 +12,8 @@
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "test": "jest --config=jest.config.json",
-    "test:coverage": "jest --config=jest.config.json --coverage",
+    "test": "npm run test:unit && npm run test:typescript",
+    "test:unit": "tap",
     "test:typescript": "tsd",
     "lint": "eslint \"lib/**/*.js\" \"test/**/*.js\" index.js",
     "lint:everything": "npm run lint && npm run test:typescript",
@@ -30,16 +30,16 @@
     "toad-scheduler": "^1.0.2"
   },
   "devDependencies": {
-    "@types/jest": "^29.0.0",
     "@types/node": "^18.0.0",
+    "@types/sinonjs__fake-timers": "^8.1.2",
     "eslint": "^8.0.0",
     "eslint-config-prettier": "^8.0.0",
     "eslint-plugin-prettier": "^4.0.0",
     "fastify": "^4.0.0-rc.3",
     "husky": "^8.0.0",
-    "jest": "^29.0.1",
     "pinst": "^3.0.0",
     "prettier": "^2.2.1",
+    "tap": "^16.3.0",
     "toad-scheduler": "^1.0.2",
     "tsd": "^0.24.1"
   },
